cyclic_fixed_queue: a more sane pop_and_push
This addresses an issue in the design of cyclic_fixed_queue.
Before, pop_and_push would take the item to push by reference. An issue arises when
the referent object being pushed is also the item to pop from the queue.

This is resolved by providing only a single overload of pop_and_push that accepts
the object to push by value, which forces either a copy or move at the call site
instead.